move few methods from main to utils
