+++ utils.py
@@ -1,6 +1,8 @@
 import json
 from pathlib import Path
 from enum import Enum
+import importlib.util
+from syftbox.lib import Client
 
 
 class ParticipantStateCols(Enum):
@@ -10,6 +12,10 @@ class ParticipantStateCols(Enum):
     ADDED_PRIVATE_DATA = "Added Private Data"
     ROUND = "Round (current/total)"
     MODEL_TRAINING_PROGRESS = "Training Progress"
+
+
+def is_dir_empty(directory: Path):
+    return not any(directory.iterdir())
 
 
 def read_json(data_path: Path):
@@ -56,3 +62,61 @@ def update_json(
             participant[column_name.value] = column_val
 
     save_json(participant_history, data_path)
+
+
+def load_model_class(model_path: Path, model_class_name: str) -> type:
+    spec = importlib.util.spec_from_file_location(model_path.stem, model_path)
+    model_arch = importlib.util.module_from_spec(spec)
+    spec.loader.exec_module(model_arch)
+    model_class = getattr(model_arch, model_class_name)
+
+    return model_class
+
+
+def get_all_directories(path: Path) -> list:
+    """
+    Returns the list of directories present in the given path
+    """
+    return [x for x in path.iterdir() if x.is_dir()]
+
+
+def get_network_participants(client: Client):
+    exclude_dir = ["apps", ".syft"]
+    entries = client.datasites.iterdir()
+
+    users = []
+    for entry in entries:
+        if entry.is_dir() and entry not in exclude_dir:
+            users.append(entry.name)
+
+    return users
+
+
+def validate_launch_config(fl_config: Path) -> bool:
+    """
+    Validates the `fl_config.json` file
+    """
+
+    try:
+        fl_config = read_json(fl_config)
+    except json.JSONDecodeError:
+        raise ValueError(f"Invalid JSON format in {fl_config.resolve()}")
+
+    required_keys = [
+        "project_name",
+        "aggregator",
+        "participants",
+        "model_arch",
+        "model_weight",
+        "model_class_name",
+        "rounds",
+        "epoch",
+        "test_dataset",
+        "learning_rate",
+    ]
+
+    for key in required_keys:
+        if key not in fl_config:
+            raise ValueError(f"Required key {key} is missing in fl_config.json")
+
+    return True
